[qwen3guard-test] relabel the OpenVINO unquantized precision fp32 -> fp16
optimum-intel's export stores the unquantized model's weights as fp16
(compress_to_fp16, the default), so the OpenVINO "fp32" artifact was
never true fp32 — its verdict logits sit ~1% off a true-fp32 reference.
Rename the precision to fp16 across the export script, the CPU run
script, the summary table, and the default-precision map so the label
matches what the runtime executes.

+++ src/gen_backends.py
@@ -7,7 +7,7 @@ that executes the forward:
 
   pytorch   — transformers AutoModelForCausalLM, float32 eager
   onnx      — ONNX Runtime, CPUExecutionProvider (fp32 / int8 / int4 weights)
-  openvino  — OpenVINO via optimum-intel (fp32 / int8 / int4 weight compression)
+  openvino  — OpenVINO via optimum-intel (fp16 / int8 / int4 weight compression)
   llamacpp  — llama.cpp via llama-cpp-python (GGUF f16 / q8_0 / q4_k_m)
 
 The exported artifacts are produced by scripts/export_gen_{onnx,openvino,gguf}.py.
@@ -31,7 +31,7 @@ L0_MAX_NEW_TOKENS = 32
 DEFAULT_PRECISION = {
     "pytorch": "fp32",
     "onnx": "fp32",
-    "openvino": "fp32",
+    "openvino": "fp16",
     "llamacpp": "f16",
 }
